refactoring some functions to util
